Move deprecated objects out of data module into dedicated mixin
