+++ cf/data/mixin/__init__.py
@@ -1,1 +1,2 @@
 from .compressedarray import CompressedArray
+from .deprecations import DataClassDeprecationsMixin
